+++ Changelog.md
@@ -19,6 +19,7 @@
 - API: New IBlockMine interface for blocks that are block mines
 - API: New method IExplosive#explodesWhenInteractedWith as well as two utility methods
 - API: New method IPasscodeProtected#savePasscodeAndSalt for more conveniently saving passcode and salt key to NBT
+- API: New methods IPasscodeProtected#setSaveSalt and IPasscodeProtected#shouldSaveSalt to control storing the object's salt into its data storage
 - Fix: Module descriptions of security sea boats do not show up correctly
 - Fix: Reinforced Mud Bricks and Blackstone Slab/Stairs don't look 100% like their vanilla counterparts
 - Fix: Holding a camera monitor makes it possible to identify disguised cameras
@@ -40,6 +41,7 @@
 - Fix: Fake water does not flow the same way as regular water
 - Fix: Laser and inventory scanner fields don't show up properly in the overlay of Jade/TOP/etc.
 - Fix: The color chooser doesn't show color
+- Fix: Loading a structure with passcode-protected objects onto itself using a structure block invalidates all passcodes
 - Removed: Configuration "ableToBreakMines"
 
 [1] Note: Frame blocks that already existed in the world prior to this update will lose their owner. These frames can be broken by anyone so they can be placed down with the proper owner again.